requirements.txt
dill==0.3.5.1
fastapi==0.78.0
flake8==4.0.1
networkx==2.8.6
PyGithub==1.55
pyjson5==1.6.1
python-dotenv==0.20.0
uvicorn==0.18.2
